requirements.txt
google-genai
opencv-python
pyaudio
mss